[go_router] Factor out test helpers from go_router_test (#2312)
* Factor out test helpers from go_router_test

* bump

* remove TODO

* un-bump

* trigger checks
